requirements.txt
gradio>=4.0.0
numpy>=1.21.0
Pillow>=9.0.0
matplotlib>=3.4.0
opencv-python>=4.5.0
torch>=2.0.0
transformers>=4.30.0